`instamatic.typing_`: rename to `_typing`
