+++ src/instamatic/microscope/components/stage.py
@@ -6,9 +6,9 @@ from typing import Optional, Tuple
 
 import numpy as np
 
+from instamatic._typing import float_deg, int_nm
 from instamatic.microscope.base import MicroscopeBase
 from instamatic.microscope.utils import StagePositionTuple
-from instamatic.typing_ import float_deg, int_nm
 
 
 class Stage:
