+++ src/instamatic/microscope/interface/fei_microscope.py
@@ -9,10 +9,10 @@ import comtypes.client
 from numpy import pi
 
 from instamatic import config
+from instamatic._typing import float_deg, int_nm
 from instamatic.exceptions import FEIValueError, TEMCommunicationError
 from instamatic.microscope.base import MicroscopeBase
 from instamatic.microscope.utils import StagePositionTuple
-from instamatic.typing_ import float_deg, int_nm
 
 logger = logging.getLogger(__name__)
