+++ src/instamatic/microscope/interface/jeol_microscope.py
@@ -8,10 +8,10 @@ from typing import Optional, Tuple
 import comtypes.client
 
 from instamatic import config
+from instamatic._typing import float_deg, int_nm
 from instamatic.exceptions import JEOLValueError, TEMCommunicationError, TEMValueError
 from instamatic.microscope.base import MicroscopeBase
 from instamatic.microscope.utils import StagePositionTuple
-from instamatic.typing_ import float_deg, int_nm
 
 logger = logging.getLogger(__name__)
