+++ src/instamatic/microscope/utils.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import NamedTuple
 
-from instamatic.typing_ import float_deg, int_nm
+from instamatic._typing import float_deg, int_nm
 
 
 class StagePositionTuple(NamedTuple):
